Move Hybrid mHC wrapper into layers module
